[WIP] Edit lynx config
For some reason, settings set from the options menu (invoked
with `o' in lynx) are not persistent. They should be saved to
~/.lynxrc, but that doesn't work with ~/.lynxrc being a symlink;
it only works when ~/.lynxrc is a regular file.

For now, if you want to change an option from the menu and want it
to persist, and you don't know the syntax to change it directly in
~/.lynxrc, perform the following process:

1. Remove the ~/.lynxrc symlink
2. Copy ~/stow-dotfiles/lynx/.lynxrc to ~/.lynxrc
3. Invoke lynx
4. Open the options menu from lynx (`o')
5. Set the options you want; be sure to check the "Save options to
   disk" checkbox before accepting the changes
6. Copy ~/.lynxrc to ~/stow-dotfiles/lynx/.lynxrc, overriding the
   existing file
7. Remove the ~/.lynxrc file
8. Symlink ~/stow-dotfiles/lynx/.lynxrc to ~/.lynxrc (e.g. via
   `make lynx' in the stow-dotfiles repository root)
9. Commit your changes to stow-dotfiles